+++ example/a.ts
@@ -1,13 +1,4 @@
 import { Elysia, t } from '../src'
 import { req } from '../test/utils'
 
-const app = new Elysia().get('/', 'Static Content', {
-	beforeHandle() {
-		return 'beforeHandle'
-	}
-})
-.listen(3000)
-
-const response = await app.handle(req('/')).then((x) => x.text())
-
-console.log(app.router.static.http.handlers[0])
+const plugin = new Elysia().get('/plugin', 'static ')